Add CreateUser

+++ src/main/java/org/Boton/model/User.java
@@ -10,17 +10,17 @@ public class User {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
     private String user_first_name;
-    private int user_telegram_id;
+    private long user_telegram_id;
     private long user_chat_id;
     private int available_means;
 
-    public User(String user_first_name, int user_telegram_id, long user_chat_id) {
+    public User(String user_first_name, long user_telegram_id, long user_chat_id) {
         this.user_first_name = user_first_name;
         this.user_telegram_id = user_telegram_id;
         this.user_chat_id = user_chat_id;
     }
 
-    public User(int user_telegram_id, int available_means) {
+    public User(long user_telegram_id, int available_means) {
         this.user_telegram_id = user_telegram_id;
         this.available_means = available_means;
     }
@@ -37,7 +37,7 @@ public class User {
         this.user_first_name = user_first_name;
     }
 
-    public int getUser_telegram_id() {
+    public long getUser_telegram_id() {
         return user_telegram_id;
     }
